+++ src/input/TestCase.java
@@ -3,9 +3,9 @@ package input;
 public abstract class TestCase {
 	public abstract int getInputSize();
 
-	public abstract double getInput(int ind);
+	public abstract float getInput(int ind);
 
 	public abstract int getOutputSize();
 
-	public abstract double getOutput(int ind);
+	public abstract float getOutput(int ind);
 }
