+++ src/math/MathUtils.java
@@ -4,11 +4,11 @@ import static java.lang.Math.exp;
 
 public class MathUtils {
 
-	public static double sigmoid(double val) {
-		return (1.0 / (1.0 + exp(-val)));
+	public static float sigmoid(double val) {
+		return (1.0f / (1.0f + (float) exp(-val)));
 	}
 
-	public static double derivative_sigmoid(double val) {
+	public static float derivative_sigmoid(double val) {
 		return sigmoid(val) * (1 - sigmoid(val));
 	}
